Update utils test.
